requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.12
albumentations>=1.3.1
opencv-python>=4.8.0
numpy>=1.23
pandas>=1.5
scikit-learn>=1.2
scipy>=1.9
matplotlib>=3.7
seaborn>=0.12
Pillow>=9.5
streamlit>=1.30
